+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### v1.1.2
+- Fixed a bug where the MathHub directory could never be changed during setup
+- Compatible with MMT 24.1.0
+- Fixed README
+- Added simple HTML export
+
 ### v1.1.1
 - Fixed a bug where spaces in the paths to executables (e.g. the jre) caused issues